requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
pyaudio
py3-tts
SpeechRecognition
python-dotenv
PyObjC==10.3.2
numpy
transformers
torch
simpleaudio
cffi==1.17.1
onnxruntime==1.20.1
piper_phonemize_cross==1.2.1
piper-tts==1.2.0
pycparser==2.22
setuptools==75.6.0
sounddevice==0.5.1
